Remove/ignore unused vars, use unsigned long long to iterate over sizes

+++ src/libsodium/crypto_stream/salsa2012/ref/stream_salsa2012.c
@@ -21,7 +21,7 @@ int crypto_stream(
 {
   unsigned char in[16];
   unsigned char block[64];
-  int i;
+  unsigned long long i;
   unsigned int u;
 
   if (!clen) return 0;
